app/web_research.py
import aiohttp
import asyncio
import logging
from typing import List, Dict, Any, Optional
from urllib.parse import urlparse
import re

from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)

class WebResearcher:
    def __init__(self):
        self.session = None
        self.headers = {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
        }
        
    async def initialize(self):
        """تهيئة باحث الويب"""
        self.session = aiohttp.ClientSession(headers=self.headers)
        logger.info("Web Researcher initialized successfully")
        
    async def search(self, query: str, max_results: int = 5) -> List[Dict]:
        """البحث على الإنترنت"""
        # في الإنتاج، استخدم API بحث مثل Google Custom Search أو SerpAPI
        # هذا تنفيذ أساسي للتوضيح
        
        try:
            # محاكاة نتائج البحث (ستحتاج إلى استبدال هذا بـ API حقيقي)
            mock_results = [
                {
                    "title": f"نتيجة بحث عن: {query}",
                    "url": f"https://example.com/{query.replace(' ', '_')}",
                    "snippet": f"هذا هو نتيجة البحث عن {query}. في التنفيذ الحقيقي، سيتم الحصول على نتائج حقيقية من API البحث."
                }
            ]
            
            return mock_results * max_results
            
        except Exception as e:
            logger.error(f"Search failed: {e}")
            return []
    
    async def extract_content(self, url: str) -> str:
        """استخراج المحتوى من URL"""
        if not self.session:
            await self.initialize()
            
        try:
            async with self.session.get(url) as response:
                if response.status == 200:
                    html = await response.text()
                    soup = BeautifulSoup(html, 'lxml')
                    
                    # إزالة العناصر غير المرغوب فيها
                    for element in soup(["script", "style", "nav", "footer", "header"]):
                        element.decompose()
                    
                    # استخراج النص الرئيسي
                    text = soup.get_text()
                    
                    # تنظيف النص
                    lines = (line.strip() for line in text.splitlines())
                    chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
                    text = ' '.join(chunk for chunk in chunks if chunk)
                    
                    return text
                else:
                    logger.error(f"Failed to fetch URL: {url}, Status: {response.status}")
                    return ""
        except Exception as e:
            logger.error(f"Failed to extract content from {url}: {e}")
            return ""
    
    async def close(self):
        """إغلاق الجلسة"""
        if self.session:
            await self.session.close()
            logger.info("Web Researcher session closed")