requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==0.28.0
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
numpy==1.24.3
pandas==2.0.3
jupyter==1.0.0
ipython==8.17.2
pinecone-client==2.2.4
chromadb==0.4.15
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.349
tiktoken==0.5.2